building_segmentation/requirements.txt
torch
torchvision
torchaudio
numpy
imageio
Pillow
einops
timm==0.4.12
fvcore
triton
